refactor: move auth functions into auth modules
